@@ -43,6 +43,7 @@ const NAMES = {
   unknown: 'unknown property type',
   willClearRender: 'lifecycle hook',
   willDestroyElement: 'lifecycle hook',
+  willInsertElement: 'lifecycle hook',
   willRender: 'lifecycle hook',
   willUpdate: 'lifecycle hook'
 };
@@ -54,6 +55,10 @@ function determinePropertyType(node, parentType) {
 
   if (ember.isInjectedControllerProp(node.value)) {
     return 'controller';
+  }
+
+  if (node.key.name === 'init') {
+    return 'init';
   }
 
   if (parentType === 'component') {
@@ -184,8 +189,60 @@ function reportUnorderedProperties(node, context, parentType, ORDER) {
       const nextTypeName = getName(firstPropertyOfNextType.type, firstPropertyOfNextType.node);
       const nextSourceLine = firstPropertyOfNextType.node.loc.start.line;
 
-      context.report(property,
-        `The ${typeName} should be above the ${nextTypeName} on line ${nextSourceLine}`);
+      context.report({
+        node: property,
+        message: `The ${typeName} should be above the ${nextTypeName} on line ${nextSourceLine}`,
+        fix: (fixer) => {
+          // for capturing the moved property and EOL character to insert ',' in between
+          const propertyWithEOL = /(.+)(\s+)$/;
+          const sourceCode = context.getSourceCode();
+          const foundProperty = firstPropertyOfNextType.node;
+          let nextToken = property;
+          let optionalComma = '';
+          let isLastProperty = false;
+
+          // including EOL character(s)
+          do {
+            const previousToken = nextToken;
+            nextToken = sourceCode.getTokenAfter(nextToken, { includeComments: true });
+
+            if (!nextToken) {
+              nextToken = previousToken;
+            }
+
+            // adding a trailing comma when it's the last property defined
+            if (nextToken.value === '}') {
+              isLastProperty = true;
+
+              if (previousToken.value !== ',') {
+                optionalComma = ',';
+              }
+            }
+          } while (nextToken.value === ',');
+
+          // additional whitespace is needed only when it's the last property
+          const whitespaceCount = isLastProperty ? property.loc.start.column : 0;
+
+          const propertyOffset = getCommentOffsetBefore(property, sourceCode);
+          const foundPropertyOffset = getCommentOffsetBefore(foundProperty, sourceCode);
+
+          const offset = nextToken.start - property.end;
+          const textBetween = property.start - propertyOffset - foundProperty.end - whitespaceCount;
+
+          const replaceTextRange = [foundProperty.start - foundPropertyOffset, nextToken.start];
+          const movedProperty = sourceCode.getText(property, propertyOffset, offset);
+          const restText = sourceCode.getText(foundProperty, foundPropertyOffset, textBetween);
+
+          // adding the optional comma between the property and newline
+          const replaceWithComma = `$1${optionalComma}$2`;
+          const movedPropertyWithComma = movedProperty.replace(propertyWithEOL, replaceWithComma);
+          const optionalWhitespace = (new Array(whitespaceCount + 1)).join(' ');
+
+          const outputText = movedPropertyWithComma + optionalWhitespace + restText;
+
+          return fixer.replaceTextRange(replaceTextRange, outputText);
+        }
+      });
     } else {
       maxOrder = order;
 
@@ -194,6 +251,23 @@ function reportUnorderedProperties(node, context, parentType, ORDER) {
       }
     }
   });
+}
+
+function getCommentOffsetBefore(property, sourceCode) {
+  const commentBlockRegExp = /^\/\*(.|\s)*\*\/$/m;
+  const commentLineRegExp = /^\/\/.*$/;
+  const previousToken = sourceCode.getTokenBefore(property, { includeComments: true });
+  const previousTokenText = sourceCode.getText(previousToken);
+
+  // including comments above the moved property
+  const isLineComment = previousToken.type === 'Line' && commentLineRegExp.test(previousTokenText);
+  const isBlockComment = previousToken.type === 'Block' && commentBlockRegExp.test(previousTokenText);
+
+  if (isLineComment || isBlockComment) {
+    return property.start - previousToken.start;
+  }
+
+  return 0;
 }
 
 function addBackwardsPosition(order, newPosition, targetPosition) {
