@@ -2,11 +2,8 @@ module.exports = {
   root: true,
 
   parserOptions: {
-    ecmaVersion: 6,
-    sourceType: 'module',
-    ecmaFeatures: {
-      experimentalObjectRestSpread: true,
-    },
+    ecmaVersion: 2018,
+    sourceType: 'module'
   },
 
   env: {
