@@ -1,20 +1,52 @@
 'use strict';
 
-const resolve = require('path').resolve;
-const requireFolderTree = require('require-folder-tree');
-
-const rules = requireFolderTree(resolve(__dirname, 'rules'));
-const configs = requireFolderTree(resolve(__dirname, 'config'));
-
-/* eslint-disable import/no-dynamic-require */
-const ember = require(resolve(__dirname, 'utils/ember'));
-const utils = require(resolve(__dirname, 'utils/utils'));
-
+/* eslint-disable global-require */
 module.exports = {
-  rules,
-  configs,
+  rules: {
+    'alias-model-in-controller': require('./rules/alias-model-in-controller'),
+    'avoid-leaking-state-in-components': require('./rules/avoid-leaking-state-in-components'),
+    'avoid-leaking-state-in-ember-objects': require('./rules/avoid-leaking-state-in-ember-objects'),
+    'avoid-using-needs-in-controllers': require('./rules/avoid-using-needs-in-controllers'),
+    'closure-actions': require('./rules/closure-actions'),
+    'jquery-ember-run': require('./rules/jquery-ember-run'),
+    'local-modules': require('./rules/local-modules'),
+    'named-functions-in-promises': require('./rules/named-functions-in-promises'),
+    'new-module-imports': require('./rules/new-module-imports'),
+    'no-attrs-in-components': require('./rules/no-attrs-in-components'),
+    'no-attrs-snapshot': require('./rules/no-attrs-snapshot'),
+    'no-capital-letters-in-routes': require('./rules/no-capital-letters-in-routes'),
+    'no-deeply-nested-dependent-keys-with-each': require('./rules/no-deeply-nested-dependent-keys-with-each'),
+    'no-duplicate-dependent-keys': require('./rules/no-duplicate-dependent-keys'),
+    'no-ember-super-in-es-classes': require('./rules/no-ember-super-in-es-classes'),
+    'no-ember-testing-in-module-scope': require('./rules/no-ember-testing-in-module-scope'),
+    'no-empty-attrs': require('./rules/no-empty-attrs'),
+    'no-function-prototype-extensions': require('./rules/no-function-prototype-extensions'),
+    'no-global-jquery': require('./rules/no-global-jquery'),
+    'no-invalid-debug-function-arguments': require('./rules/no-invalid-debug-function-arguments'),
+    'no-jquery': require('./rules/no-jquery'),
+    'no-new-mixins': require('./rules/no-new-mixins'),
+    'no-observers': require('./rules/no-observers'),
+    'no-old-shims': require('./rules/no-old-shims'),
+    'no-on-calls-in-components': require('./rules/no-on-calls-in-components'),
+    'no-restricted-resolver-tests': require('./rules/no-restricted-resolver-tests'),
+    'no-side-effects': require('./rules/no-side-effects'),
+    'no-test-and-then': require('./rules/no-test-and-then'),
+    'no-test-import-export': require('./rules/no-test-import-export'),
+    'order-in-components': require('./rules/order-in-components'),
+    'order-in-controllers': require('./rules/order-in-controllers'),
+    'order-in-models': require('./rules/order-in-models'),
+    'order-in-routes': require('./rules/order-in-routes'),
+    'require-super-in-init': require('./rules/require-super-in-init'),
+    'routes-segments-snake-case': require('./rules/routes-segments-snake-case'),
+    'use-brace-expansion': require('./rules/use-brace-expansion'),
+    'use-ember-get-and-set': require('./rules/use-ember-get-and-set'),
+  },
+  configs: {
+    base: require('./config/base'),
+    recommended: require('./config/recommended'),
+  },
   utils: {
-    ember,
-    utils,
+    ember: require('./utils/ember'),
+    utils: require('./utils/utils'),
   },
 };
