@@ -8,6 +8,7 @@ module.exports = {
   "ember/alias-model-in-controller": "off",
   "ember/avoid-leaking-state-in-components": "off",
   "ember/avoid-leaking-state-in-ember-objects": "error",
+  "ember/avoid-using-needs-in-controllers": "error",
   "ember/closure-actions": "error",
   "ember/jquery-ember-run": "error",
   "ember/local-modules": "off",
@@ -16,15 +17,23 @@ module.exports = {
   "ember/no-attrs-in-components": "error",
   "ember/no-attrs-snapshot": "error",
   "ember/no-capital-letters-in-routes": "error",
+  "ember/no-deeply-nested-dependent-keys-with-each": "off",
   "ember/no-duplicate-dependent-keys": "error",
+  "ember/no-ember-super-in-es-classes": "off",
+  "ember/no-ember-testing-in-module-scope": "error",
   "ember/no-empty-attrs": "off",
   "ember/no-function-prototype-extensions": "error",
   "ember/no-global-jquery": "error",
+  "ember/no-invalid-debug-function-arguments": "off",
   "ember/no-jquery": "off",
-  "ember/no-observers": "off",
+  "ember/no-new-mixins": "off",
+  "ember/no-observers": "error",
   "ember/no-old-shims": "error",
   "ember/no-on-calls-in-components": "error",
+  "ember/no-restricted-resolver-tests": "error",
   "ember/no-side-effects": "error",
+  "ember/no-test-and-then": "off",
+  "ember/no-test-import-export": "off",
   "ember/order-in-components": "off",
   "ember/order-in-controllers": "off",
   "ember/order-in-models": "off",
