@@ -12,7 +12,8 @@ module.exports = {
     docs: {
       description: 'Enforces aliasing model in controller',
       category: 'Best Practices',
-      recommended: false
+      recommended: false,
+      url: 'https://github.com/ember-cli/eslint-plugin-ember/tree/master/docs/rules/alias-model-in-controller.md'
     },
     fixable: null, // or "code" or "whitespace"
   },
