@@ -39,7 +39,8 @@ module.exports = {
     docs: {
       description: 'Avoids state leakage',
       category: 'Ember Object',
-      recommended: true
+      recommended: true,
+      url: 'https://github.com/ember-cli/eslint-plugin-ember/tree/master/docs/rules/avoid-leaking-state-in-ember-objects.md'
     },
     fixable: null, // or "code" or "whitespace"
     schema: [{
@@ -47,6 +48,8 @@ module.exports = {
       items: { type: 'string' },
     }],
   },
+
+  DEFAULT_IGNORED_PROPERTIES,
 
   create(context) {
     const ignoredProperties = context.options[0] || DEFAULT_IGNORED_PROPERTIES;
@@ -63,7 +66,9 @@ module.exports = {
         const properties = ember.getModuleProperties(node);
 
         properties
-          .filter(property => ignoredProperties.indexOf(property.key.name) === -1)
+          .filter(property =>
+            property.key && ignoredProperties.indexOf(property.key.name) === -1
+          )
           .forEach((property) => {
             if (!isAllowed(property)) {
               report(property);
