@@ -11,7 +11,8 @@ module.exports = {
     docs: {
       description: 'Enforces usage of closure actions',
       category: 'Best Practices',
-      recommended: true
+      recommended: true,
+      url: 'https://github.com/ember-cli/eslint-plugin-ember/tree/master/docs/rules/closure-actions.md'
     },
     fixable: null, // or "code" or "whitespace"
   },
