@@ -12,7 +12,8 @@ module.exports = {
     docs: {
       description: 'Prevents usage of jQuery without Ember Run Loop',
       category: 'Possible Errors',
-      recommended: true
+      recommended: true,
+      url: 'https://github.com/ember-cli/eslint-plugin-ember/tree/master/docs/rules/jquery-ember-run.md'
     },
     fixable: null, // or "code" or "whitespace"
   },
