@@ -11,7 +11,8 @@ module.exports = {
     docs: {
       description: 'Enforces usage of named functions in promises',
       category: 'Best Practices',
-      recommended: false
+      recommended: false,
+      url: 'https://github.com/ember-cli/eslint-plugin-ember/tree/master/docs/rules/named-functions-in-promises.md'
     },
     fixable: null, // or "code" or "whitespace"
   },
