@@ -1,7 +1,21 @@
 'use strict';
 
 const EMBER_NAMESPACES = ['inject.controller', 'inject.service'];
-const GLOBALS = require('ember-rfc176-data/globals.json');
+const MAPPING = require('ember-rfc176-data');
+
+const GLOBALS = MAPPING.reduce((memo, exportDefinition) => {
+  if (exportDefinition.deprecated) {
+    return memo;
+  }
+
+  if (exportDefinition.global in memo) {
+    return memo;
+  }
+
+  memo[exportDefinition.global] = exportDefinition; // eslint-disable-line no-param-reassign
+
+  return memo;
+}, Object.create(null));
 
 //------------------------------------------------------------------------------
 // General rule - Use "New Module Imports" from Ember RFC #176
@@ -12,7 +26,8 @@ module.exports = {
     docs: {
       description: ' Use "New Module Imports" from Ember RFC #176',
       category: 'Best Practices',
-      recommended: true
+      recommended: true,
+      url: 'https://github.com/ember-cli/eslint-plugin-ember/tree/master/docs/rules/new-module-imports.md'
     },
   },
 
@@ -32,7 +47,7 @@ module.exports = {
             reportNestedProperties(props, parent);
           } else {
             const key = item.key.name;
-            const match = GLOBALS[key];
+            const match = GLOBALS[`Ember.${key}`];
             populateMessage({
               node: item,
               customKey: (key !== item.value.name) ? item.value.name : null,
@@ -66,7 +81,7 @@ module.exports = {
           fullName = fullNames[i];
 
           const key = fullName.replace(/^Ember\./, '');
-          const match = GLOBALS[key];
+          const match = GLOBALS[fullName];
 
           const reportedError = populateMessage({ node, fullName, key, match });
 
@@ -80,7 +95,7 @@ module.exports = {
 
     function reportNestedProperties(properties, parent) {
       properties.forEach((item) => {
-        const match = GLOBALS[`${parent}.${item.key.name}`];
+        const match = GLOBALS[`Ember.${parent}.${item.key.name}`];
 
         populateMessage({
           node: item,
@@ -107,18 +122,20 @@ module.exports = {
       // Ex: import { inject as service } from '@ember/service';
       const isNamespace = EMBER_NAMESPACES.indexOf(`${obj.parent}.${obj.key}`) !== -1;
 
+      const isNamedExport = obj.match.export !== 'default';
+
       let importSpecifier;
       let message;
 
-      if (obj.match[1] && !isNamespace) {
-        importSpecifier = obj.match[2] ? `{ ${obj.match[1]} as ${obj.match[2]} }` : `{ ${obj.match[1]} }`;
-      } else if (obj.match[1] && isNamespace) {
+      if (isNamedExport && !isNamespace) {
+        importSpecifier = obj.match.localName ? `{ ${obj.match.export} as ${obj.match.localName} }` : `{ ${obj.match.export} }`;
+      } else if (isNamedExport && isNamespace) {
         importSpecifier = `{ ${obj.parent} as ${obj.key} }`;
       } else {
-        importSpecifier = obj.match[2] || obj.customKey || obj.key;
+        importSpecifier = obj.match.localName || obj.customKey || obj.key;
       }
 
-      const replacement = `import ${importSpecifier} from '${obj.match[0]}';`;
+      const replacement = `import ${importSpecifier} from '${obj.match.module}';`;
 
       if (obj.type === 'Property') {
         message = `Use ${replacement} instead of using Ember destructuring`;
