@@ -13,7 +13,8 @@ module.exports = {
     docs: {
       description: 'Disallow usage of this.attrs in components',
       category: 'Possible Errors',
-      recommended: true
+      recommended: true,
+      url: 'https://github.com/ember-cli/eslint-plugin-ember/tree/master/docs/rules/no-attrs-in-components.md'
     },
     fixable: null,
   },
