@@ -14,7 +14,8 @@ module.exports = {
     docs: {
       description: 'Disallow use of attrs snapshot in `didReceiveAttrs` and `didUpdateAttrs`',
       category: 'Possible Errors',
-      recommended: true
+      recommended: true,
+      url: 'https://github.com/ember-cli/eslint-plugin-ember/tree/master/docs/rules/no-attrs-snapshot.md'
     },
     fixable: null, // or "code" or "whitespace"
     message
