@@ -1,6 +1,8 @@
 'use strict';
 
 const ember = require('../utils/ember');
+const utils = require('../utils/utils');
+
 //------------------------------------------------------------------------------
 // Routing - No capital letters in routes
 //------------------------------------------------------------------------------
@@ -11,6 +13,7 @@ module.exports = {
       description: 'Raise an error when there is a route with uppercased letters in router.js',
       category: 'Possible Errors',
       recommended: true,
+      url: 'https://github.com/ember-cli/eslint-plugin-ember/tree/master/docs/rules/no-capital-letters-in-routes.md'
     },
     fixable: null, // or "code" or "whitespace"
     schema: [],
@@ -22,19 +25,13 @@ module.exports = {
 
     return {
       CallExpression(node) {
-        if (
-          !ember.isRoute(node) ||
-          !node.arguments[0] ||
-          (node.arguments[0] && node.arguments[0].type === 'Identifier')
-        ) {
-          return;
-        }
+        if (ember.isRoute(node) && node.arguments[0] && utils.isLiteral(node.arguments[0])) {
+          const routeName = node.arguments[0].value;
+          const hasAnyUppercaseLetter = Boolean(routeName.match('[A-Z]'));
 
-        const routeName = node.arguments[0].value;
-        const hasAnyUppercaseLetter = Boolean(routeName.match('[A-Z]'));
-
-        if (hasAnyUppercaseLetter) {
-          report(node);
+          if (hasAnyUppercaseLetter) {
+            report(node);
+          }
         }
       },
     };
