@@ -12,7 +12,8 @@ module.exports = {
     docs: {
       description: 'Disallow repeating dependent keys',
       category: 'Possible Errors',
-      recommended: true
+      recommended: true,
+      url: 'https://github.com/ember-cli/eslint-plugin-ember/tree/master/docs/rules/no-duplicate-dependent-keys.md'
     },
     fixable: null,
     message: MESSAGE,
