@@ -11,7 +11,8 @@ module.exports = {
     docs: {
       description: 'Prevents usage of empty attributes in ember data models',
       category: 'Best Practices',
-      recommended: false
+      recommended: false,
+      url: 'https://github.com/ember-cli/eslint-plugin-ember/tree/master/docs/rules/no-empty-attrs.md'
     },
     fixable: null, // or "code" or "whitespace"
   },
