@@ -11,7 +11,8 @@ module.exports = {
     docs: {
       description: 'Prevents usage of Ember\'s `function` prototype extensions',
       category: 'Best Practices',
-      recommended: true
+      recommended: true,
+      url: 'https://github.com/ember-cli/eslint-plugin-ember/tree/master/docs/rules/no-function-prototype-extensions.md'
     },
     fixable: null, // or "code" or "whitespace"
   },
@@ -19,7 +20,7 @@ module.exports = {
   create(context) {
     const message = 'Don\'t use Ember\'s function prototype extensions';
 
-    const functionPrototypeExtensionNames = ['property', 'observe', 'on'];
+    const functionPrototypeExtensionNames = ['property', 'observes', 'on'];
 
     const isFunctionPrototypeExtension = function (property) {
       return utils.isIdentifier(property) &&
