@@ -16,8 +16,9 @@ module.exports = {
       description: 'Prevents usage of global jQuery object',
       category: 'Best Practices',
       recommended: true,
+      url: 'https://github.com/ember-cli/eslint-plugin-ember/tree/master/docs/rules/no-global-jquery.md'
     },
-    fixable: null,  // or "code" or "whitespace"
+    fixable: null, // or "code" or "whitespace"
     message: MESSAGE,
   },
 
