@@ -15,7 +15,8 @@ module.exports = {
     docs: {
       description: 'Disallow any usage of jQuery',
       category: 'Best Practices',
-      recommended: false
+      recommended: false,
+      url: 'https://github.com/ember-cli/eslint-plugin-ember/tree/master/docs/rules/no-jquery.md'
     },
     fixable: null, // or "code" or "whitespace"
     message
@@ -23,6 +24,7 @@ module.exports = {
   create(context) {
     let destructuredAssignment;
     let emberImportAliasName;
+    let jqueryImportAliasName;
     const report = function (node) {
       context.report(node, message);
     };
@@ -30,6 +32,10 @@ module.exports = {
     return {
       ImportDeclaration(node) {
         emberImportAliasName = ember.getEmberImportAliasName(node);
+
+        if (!jqueryImportAliasName && node.source && node.source.value === 'jquery') {
+          jqueryImportAliasName = node.specifiers[0].local.name;
+        }
       },
 
       VariableDeclarator(node) {
@@ -50,6 +56,10 @@ module.exports = {
           node.object.name === 'Em' ||
           (emberImportAliasName && node.object.name === emberImportAliasName)) &&
           node.property.name === '$') {
+          report(node);
+        }
+
+        if (jqueryImportAliasName && node.object.name === jqueryImportAliasName) {
           report(node);
         }
       },
