@@ -4,29 +4,44 @@ const ember = require('../utils/ember');
 const utils = require('../utils/utils');
 
 //----------------------------------------------
-// General rule - Don\'t use .on() in components
+// Don’t use .on() for component lifecycle events.
 //----------------------------------------------
 
 module.exports = {
   meta: {
     docs: {
-      description: 'Prevents usage of `on` calls in components',
+      description: 'Prevents usage of `on` to call lifecycle hooks in components',
       category: 'Best Practices',
-      recommended: true
+      recommended: true,
+      url: 'https://github.com/ember-cli/eslint-plugin-ember/tree/master/docs/rules/no-on-calls-in-components.md'
     },
     fixable: null, // or "code" or "whitespace"
   },
 
   create(context) {
-    const message = 'Don\'t use .on() in components';
+    const message = 'Don\'t use .on() for component lifecycle events.';
+    const lifecycleHooks = ['init',
+      'didUpdateAttrs',
+      'didReceiveAttrs',
+      'willUpdate',
+      'willRender',
+      'didInsertElement',
+      'didUpdate',
+      'didRender',
+      'willDestroyElement',
+      'willClearRender',
+      'didDestroyElement'
+    ];
 
     const filePath = context.getFilename();
 
     const isOnCall = function (node) {
       const value = node.value;
       const callee = value.callee;
+      const args = utils.parseArgs(value);
 
       return utils.isCallExpression(value) &&
+        lifecycleHooks.indexOf(args[0]) > -1 &&
         (
           (utils.isIdentifier(callee) && callee.name === 'on') ||
           (
