@@ -16,6 +16,7 @@ const ORDER = [
   'init',
   'didReceiveAttrs',
   'willRender',
+  'willInsertElement',
   'didInsertElement',
   'didRender',
   'didUpdateAttrs',
@@ -37,9 +38,10 @@ module.exports = {
     docs: {
       description: 'Enforces proper order of properties in components',
       category: 'Stylistic Issues',
-      recommended: false
+      recommended: false,
+      url: 'https://github.com/ember-cli/eslint-plugin-ember/tree/master/docs/rules/order-in-components.md'
     },
-    fixable: null // or "code" or "whitespace"
+    fixable: 'code' // or "code" or "whitespace"
   },
 
   create(context) {
@@ -53,6 +55,7 @@ module.exports = {
         'init',
         'didReceiveAttrs',
         'willRender',
+        'willInsertElement',
         'didInsertElement',
         'didRender',
         'didUpdateAttrs',
