@@ -12,6 +12,7 @@ const ORDER = [
   'query-params',
   'inherited-property',
   'property',
+  'init',
   'single-line-function',
   'multi-line-function',
   'observer',
@@ -28,9 +29,10 @@ module.exports = {
     docs: {
       description: 'Enforces proper order of properties in controllers',
       category: 'Stylistic Issues',
-      recommended: false
+      recommended: false,
+      url: 'https://github.com/ember-cli/eslint-plugin-ember/tree/master/docs/rules/order-in-controllers.md'
     },
-    fixable: null, // or "code" or "whitespace"
+    fixable: 'code', // or "code" or "whitespace"
   },
 
   create(context) {
