@@ -23,9 +23,10 @@ module.exports = {
     docs: {
       description: 'Enforces proper order of properties in models',
       category: 'Stylistic Issues',
-      recommended: false
+      recommended: false,
+      url: 'https://github.com/ember-cli/eslint-plugin-ember/tree/master/docs/rules/order-in-models.md'
     },
-    fixable: null, // or "code" or "whitespace"
+    fixable: 'code', // or "code" or "whitespace"
   },
 
   create(context) {
