@@ -10,6 +10,7 @@ const ORDER = [
   'service',
   'inherited-property',
   'property',
+  'init',
   'single-line-function',
   'multi-line-function',
   'beforeModel',
@@ -35,9 +36,10 @@ module.exports = {
     docs: {
       description: 'Enforces proper order of properties in routes',
       category: 'Stylistic Issues',
-      recommended: false
+      recommended: false,
+      url: 'https://github.com/ember-cli/eslint-plugin-ember/tree/master/docs/rules/order-in-routes.md'
     },
-    fixable: null // or "code" or "whitespace"
+    fixable: 'code' // or "code" or "whitespace"
   },
 
   create(context) {
