@@ -13,7 +13,8 @@ module.exports = {
     docs: {
       description: 'Enforces usage of snake_cased dynamic segments in routes',
       category: 'Possible Errors',
-      recommended: true
+      recommended: true,
+      url: 'https://github.com/ember-cli/eslint-plugin-ember/tree/master/docs/rules/routes-segments-snake-case.md'
     },
     fixable: null, // or "code" or "whitespace"
   },
