@@ -12,7 +12,8 @@ module.exports = {
     docs: {
       description: 'Enforces usage of brace expansion',
       category: 'Stylistic Issues',
-      recommended: true
+      recommended: true,
+      url: 'https://github.com/ember-cli/eslint-plugin-ember/tree/master/docs/rules/use-brace-expansion.md'
     },
     fixable: null, // or "code" or "whitespace"
   },
@@ -26,7 +27,9 @@ module.exports = {
 
     return {
       CallExpression(node) {
-        if (!ember.isComputedProp(node)) return;
+        if (!ember.isComputedProp(node) || utils.isMemberExpression(node.callee)) {
+          return;
+        }
 
         const matchesBraces = x => !!x.match(/[{}]/g);
         const hasBraces = arr => arr.some(matchesBraces);
