@@ -17,7 +17,8 @@ module.exports = {
     docs: {
       description: 'Enforces usage of Ember.get and Ember.set',
       category: 'Best Practices',
-      recommended: false
+      recommended: false,
+      url: 'https://github.com/ember-cli/eslint-plugin-ember/tree/master/docs/rules/use-ember-get-and-set.md'
     },
     fixable: 'code',
   },
