@@ -10,9 +10,10 @@ module.exports = {
   isEmberCoreModule,
   isEmberComponent,
   isEmberController,
-  isEmberRoute,
   isEmberMixin,
   isEmberTest,
+  isEmberRoute,
+  isEmberService,
 
   isSingleLineFn,
   isMultiLineFn,
@@ -33,16 +34,11 @@ module.exports = {
 
   isRelation,
 
-  isComponentLifecycleHookName,
   isComponentLifecycleHook,
-  isComponentCustomFunction,
 
   isRoute,
-  isRouteCustomFunction,
-  isRouteProperty,
   isRouteDefaultProp,
 
-  isControllerProperty,
   isControllerDefaultProp,
   parseDependentKeys,
   unwrapBraceExpressions,
@@ -51,6 +47,7 @@ module.exports = {
   isEmberObject,
 
   isReopenObject,
+  isReopenClassObject,
 };
 
 // Private
@@ -139,6 +136,10 @@ function isEmberTest(filePath) {
   return filePath.indexOf('-test') > -1;
 }
 
+function isReopenClassObject(node) {
+  return node.callee.property && node.callee.property.name === 'reopenClass';
+}
+
 function isReopenObject(node) {
   return node.callee.property && node.callee.property.name === 'reopen';
 }
@@ -151,12 +152,20 @@ function isEmberController(node, filePath) {
   return isEmberCoreModule(node, 'Controller', filePath);
 }
 
+function isEmberMixin(node, filePath) {
+  return isEmberCoreModule(node, 'Mixin', filePath);
+}
+
 function isEmberRoute(node, filePath) {
   return isEmberCoreModule(node, 'Route', filePath);
 }
 
 function isEmberMixin(node, filePath) {
   return isEmberCoreModule(node, 'Mixin', filePath);
+}
+
+function isEmberService(node, filePath) {
+  return isEmberCoreModule(node, 'Service', filePath);
 }
 
 function isInjectedServiceProp(node) {
@@ -222,9 +231,9 @@ function isActionsProp(property) {
 
 function isComponentLifecycleHookName(name) {
   return [
-    'init',
     'didReceiveAttrs',
     'willRender',
+    'willInsertElement',
     'didInsertElement',
     'didRender',
     'didUpdateAttrs',
@@ -239,11 +248,6 @@ function isComponentLifecycleHookName(name) {
 function isComponentLifecycleHook(property) {
   return isFunctionExpression(property.value) &&
     isComponentLifecycleHookName(property.key.name);
-}
-
-function isComponentCustomFunction(property) {
-  return isFunctionExpression(property.value) &&
-    !isComponentLifecycleHookName(property.key.name);
 }
 
 function isRoute(node) {
@@ -268,14 +272,8 @@ function isRouteLifecycleHookName(name) {
   ].indexOf(name) > -1;
 }
 
-function isRouteCustomFunction(property) {
-  return isFunctionExpression(property.value) &&
-    !isRouteLifecycleHookName(property.key.name);
-}
-
 function isRouteProperty(name) {
   return [
-    'init',
     'actions',
     'concatenatedProperties',
     'controller',
@@ -296,7 +294,6 @@ function isRouteDefaultProp(property) {
 
 function isControllerProperty(name) {
   return [
-    'init',
     'actions',
     'concatenatedProperties',
     'isDestroyed',
@@ -391,7 +388,7 @@ function hasDuplicateDependentKeys(callExp) {
  * It also unwraps the expressions, so that `model.{foo,bar}` becomes `model.foo, model.bar`.
  *
  * @param  {CallExpression} callExp CallExpression to examine
- * @return {Array}          Array of unwrapped dependent keys
+ * @return {String[]}       Array of unwrapped dependent keys
  */
 function parseDependentKeys(callExp) {
   // Check whether we have a MemberExpression, eg. computed(...).volatile()
@@ -411,14 +408,14 @@ function parseDependentKeys(callExp) {
 /**
  * Unwraps brace expressions.
  *
- * @param  {Array} dependentKeys array of strings containing unprocessed dependent keys.
- * @return {Array} Array of unwrapped dependent keys
+ * @param  {String[]} dependentKeys array of strings containing unprocessed dependent keys.
+ * @return {String[]} Array of unwrapped dependent keys
  */
 function unwrapBraceExpressions(dependentKeys) {
   const braceExpressionRegexp = /{.+}/g;
 
   const unwrappedExpressions = dependentKeys.map((key) => {
-    if (!braceExpressionRegexp.test(key)) return key;
+    if (typeof key !== 'string' || !braceExpressionRegexp.test(key)) return key;
 
     const braceExpansionPart = key.match(braceExpressionRegexp)[0];
     const prefix = key.replace(braceExpansionPart, '');
