@@ -28,14 +28,16 @@ module.exports = {
   collectObjectPatternBindings,
   isEmptyMethod,
   isBinaryExpression,
+  isReturnStatement,
+  getParent
 };
 
 /**
  * Find nodes of given name
  *
- * @param  {Array} body Array of nodes
+ * @param  {Node[]} body Array of nodes
  * @param  {String} nodeName
- * @return {Array}
+ * @return {Node[]}
  */
 function findNodes(body, nodeName) {
   let nodesArray = [];
@@ -197,6 +199,16 @@ function isCallWithFunctionExpression(node) {
 }
 
 /**
+ * Check whether or not a node is a ReturnStatement
+ *
+ * @param {Object} node The node to check.
+ * @return {Boolean} Whether or not the node is a ReturnStatement.
+ */
+function isReturnStatement(node) {
+  return node !== undefined && node.type && node.type === 'ReturnStatement';
+}
+
+/**
  * Check whether or not a node is an ThisExpression.
  *
  * @param {Object} node The node to check.
@@ -245,11 +257,11 @@ function isTemplateLiteral(node) {
  *
  * @param  {CallExpression} node The node to check.
  * @param  {String}         destructuredName The desctructured name.
- * @param  {Array}          aliases array of aliases of the global function.
+ * @param  {String[]}       aliases array of aliases of the global function.
  * @return {Boolean}        Whether function is a global call
  */
-function isGlobalCallExpression(node, desctructuredName, aliases) {
-  const isDestructured = node && node.callee && node.callee.name === desctructuredName;
+function isGlobalCallExpression(node, destructuredName, aliases) {
+  const isDestructured = node && node.callee && node.callee.name === destructuredName;
   const isGlobalCall = node.callee && aliases.indexOf(node.callee.name) > -1;
 
   return !isDestructured && isGlobalCall;
@@ -269,7 +281,7 @@ function getSize(node) {
  * Parse CallExpression or NewExpression to get array of properties and object name
  *
  * @param  {Object} node The node to parse
- * @return {Array} eg. ['Ember', 'computed', 'alias']
+ * @return {String[]} eg. ['Ember', 'computed', 'alias']
  */
 function parseCallee(node) {
   const parsedCallee = [];
@@ -297,7 +309,7 @@ function parseCallee(node) {
  * Parse CallExpression to get array of arguments
  *
  * @param  {Object} node Node to parse
- * @return {Array} Literal function's arguments
+ * @return {String[]} Literal function's arguments
  */
 function parseArgs(node) {
   let parsedArgs = [];
@@ -314,8 +326,8 @@ function parseArgs(node) {
 /**
  * Find property that is in wrong order
  *
- * @param  {Array} arr Properties with their order value
- * @return {Object}    Unordered property or null
+ * @param  {Object[]} arr Properties with their order value
+ * @return {Object}       Unordered property or null
  */
 function findUnorderedProperty(arr) {
   for (let i = 0; i < arr.length - 1; i++) {
@@ -364,7 +376,7 @@ function getPropertyValue(node, path) {
  *
  * @param  {VariableDeclarator} node node to parse
  * @param  {Object}             initialObjToBinding relevant bindings
- * @return {Array}              list of object pattern bindings
+ * @return {String[]}           list of object pattern bindings
  */
 function collectObjectPatternBindings(node, initialObjToBinding) {
   if (!isObjectPattern(node.id)) return [];
@@ -391,4 +403,26 @@ function isEmptyMethod(node) {
   return node.value.body
     && node.value.body.body
     && node.value.body.body.length <= 0;
+}
+
+/**
+ * Travels up the ancestors of a given node, if the predicate function returns
+ * truthy for a given node or ancestor, return that node, otherwise return null
+ *
+ * @name getParent
+ * @param {Object} node The child node to start at
+ * @param {Function} predicate Function that should return a boolean for a given value
+ * @returns {Object|null} The first node that matches predicate, otherwise null
+ */
+function getParent(node, predicate) {
+  let currentNode = node;
+  while (currentNode) {
+    if (predicate(currentNode)) {
+      return currentNode;
+    }
+
+    currentNode = currentNode.parent;
+  }
+
+  return null;
 }
